FIX escape & to &amp; for HTML output in view.php
This is a repost of r5354, with the double-encoding of the terms.php
url-parameter fixed. Additionally, the & before
'<?php echo hook("nextpreviousextraurl") ?>' is now escaped.

The goal of this patch is to bring view.php closer to a valid HTML5
document, so it is easier to see any remaining errors.